changed structure of the application
