+++ app.js
@@ -4,8 +4,7 @@ const path = require('path');
 const cookieParser = require('cookie-parser');
 const logger = require('morgan');
 
-const indexRouter = require('./routes/index');
-const registerRouter = require('./routes/register');
+const routes = require('./routes/index');
 
 const app = express();
 
@@ -19,8 +18,7 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', indexRouter);
-app.use('/register', registerRouter);
+app.use('/', routes)
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
