+++ controllers/registerController.js
@@ -1,0 +1,8 @@
+module.exports = {
+  index(req, res) {
+    res.render('register');
+  },
+  register(req, res) {
+    res.render('login')
+  }
+};
