+++ routes/index.js
@@ -1,21 +1,9 @@
 const express = require('express');
 const router = express.Router();
+const loginRouter = require('./loginRouter');
+const registerRouter = require('./registerRouter');
 
-/* GET Login page */
-router.get('/', function(req, res, next) {
-  res.render('index');
-});
-
-/* POST Login page  */
-router.post('/', function(req, res, next) {
-  const mailAddress = req.body['mail-address'];
-  const password = req.body['password'];
-
-  const blankError = (target) => `${target}が入力されていません`;
-  if (!mailAddress) return res.render('index', { errorMessage: blankError('メールアドレス') });
-  if (!password) return res.render('index', { errorMessage: blankError('パスワード') });
-
-  res.render('/index');
-});
+router.use('/', loginRouter)
+router.use('/register', registerRouter)
 
 module.exports = router;
